requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
idna==2.7
Markdown==2.6.11
oauthlib==2.1.0
Pillow==5.2.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
dj-database-url==0.4.2
gunicorn==19.7.1
psycopg2==2.7.4
